Refactored packages & fixed lossy conversion
