+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/ATVision.java
@@ -1,6 +1,6 @@
-package org.firstinspires.ftc.teamcode.vision2;
+package org.firstinspires.ftc.teamcode.vision;
 
-import static org.firstinspires.ftc.teamcode.vision2.VisionConstants.getCenterStageTagLibrary;
+import static org.firstinspires.ftc.teamcode.vision.VisionConstants.getCenterStageTagLibrary;
 
 import android.util.Size;
