+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/VisionConstants.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.vision2;
+package org.firstinspires.ftc.teamcode.vision;
 
 import com.acmerobotics.roadrunner.Vector2d;
