+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/pipelines/eocvsim_workspace.json
@@ -1,0 +1,7 @@
+{
+  "sourcesPath": ".",
+  "resourcesPath": ".",
+  "excludedPaths": [],
+  "excludedFileExtensions": [],
+  "eocvSimVersion": "3.5.2"
+}
